requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.215
langchainplus-sdk==0.0.17
llama-index==0.6.33
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
